add integration test for run_tensorflow_script
