+++ .github/workflows/ci.yml
@@ -62,6 +62,9 @@ jobs:
             if [[ $UNAME == "linux" ]] && [[ ! "$PYTHON" =~ "3.5" ]] && [[ ! "$PYTHON" =~ "2.7" ]]; then
               conda install --quiet --yes -n test -c conda-forge --no-deps python=$PYTHON libxgboost py-xgboost xgboost;
             fi
+            if [[ $UNAME == "linux" ]]; then
+              pip install tensorflow
+            fi
             virtualenv testenv && source testenv/bin/activate;
             pip install -r requirements.txt && pip install pytest pytest-timeout;
             if [ -z "$DEFAULT_VENV" ]; then
