+++ mars/scheduler/graph.py
@@ -603,7 +603,7 @@ class GraphActor(SchedulerActor):
         initial_chunks = [c for c in chunk_graph
                           if chunk_graph.count_successors(c) == 0]
         if all(c.op.expect_worker is not None for c in initial_chunks):
-            assignments = {c: c.op.expect_worker for c in initial_chunks}
+            assignments = {c.op.key: c.op.expect_worker for c in initial_chunks}
         else:
             if analyzer is None:
                 analyzer = GraphAnalyzer(chunk_graph, self._get_worker_slots())
